@@ -1,45 +1,43 @@
 ﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using System.IO;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
 namespace netCoreWorkshop
 {
-	class Program
-	{
-		// Minimal ASP.NET setup in Main method
-		public static void Main(string[] args)
-		{
-			CreateWebHostBuilder(args).Build().Run();
-		}
+    class Program
+    {
+        // Minimal ASP.NET setup in Main method
+        public static void Main(string[] args)
+        {
+            CreateWebHostBuilder(args).Build().Run();
+        }
 
-		public static IWebHostBuilder CreateWebHostBuilder(string[] args)
-		{
-			var commandLineConfig = new ConfigurationBuilder()  
-				.AddCommandLine(args)
-				.Build();
+        public static IWebHostBuilder CreateWebHostBuilder(string[] args)
+        {
+            var commandLineConfig = new ConfigurationBuilder()
+                .AddCommandLine(args)
+                .Build();
 
-			return WebHost.CreateDefaultBuilder(args)
-				.UseConfiguration(commandLineConfig)
-				.ConfigureAppConfiguration((builderContext, config) =>
-				{
-					IHostingEnvironment env = builderContext.HostingEnvironment;
+            return WebHost.CreateDefaultBuilder(args)
+                .UseConfiguration(commandLineConfig)
+                .ConfigureAppConfiguration((builderContext, config) =>
+                {
+                    var env = builderContext.HostingEnvironment;
 
-					config.AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-						.AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: true);
+                    config.AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+                        .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: true);
 
-					config.AddEnvironmentVariables();
-				})
-				.ConfigureLogging((hostingContext, logging) =>
-				{
-					logging.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
-					logging.AddConsole();
-					logging.AddDebug();
-				})
-				.UseStartup<Startup>();
-		}
-	}
+                    config.AddEnvironmentVariables();
+                })
+                .ConfigureLogging((hostingContext, logging) =>
+                {
+                    logging.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
+                    logging.AddConsole();
+                    logging.AddDebug();
+                })
+                .UseStartup<Startup>();
+        }
+    }
 }
